requirements.txt
httpx==0.28.1
python-dotenv==1.2.1
rich==14.2.0
typer==0.21.1
